reorganized statebatch package to different files
